laundry_guest/backend/src/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mysqlclient==1.4.6
PyJWT==1.7.1
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0